+++ Magic/CHANGELOG.md
@@ -28,6 +28,7 @@
  - Allow block/reflect parameters on classes and modifiers
  - Add reflect_chance/reflect_fov parameters for classes, wands and modifiers to reflect without blocking
  - Added Reflect level 5, which makes the caster directly reflective
+ - Added `spawn` sub-actions to EntityProjectile action
 
 ## Changes